src/requirements.txt
discord.py==2.1.0
ffmpeg==1.4
ffmpeg_python==0.2.0
requests==2.28.1
